Resize egl window when software keyboard is shown
* This only works for mobile. On other devices, needs some improvements.
For examples, wearables have a very small UI area outside the keyboard area.
The TV-emulator has a translucent software keyboard, so window resize is unnecessary.
Lastly, on real TV device, the current way that to handle HW-back key fallback
on text_input_channel is not enough.
We have to take the time and improve it one by one.

Signed-off-by: Boram Bae <boram21.bae@samsung.com>